+++ hvad/tests/admin.py
@@ -162,10 +162,13 @@ class AdminMethodsTests(HvadTestCase, BaseAdminTests, NormalFixture):
         # Check what happens if there is no translations at all
         obj = Normal.objects.untranslated().create(shared_field="shared")
         with translation.override('en'):
-            self.assertEqual(myadmin.get_object(get_request, obj.pk).pk, obj.pk)
-            self.assertEqual(myadmin.get_object(get_request, obj.pk).shared_field, obj.shared_field)
-            self.assertEqual(myadmin.get_object(get_request, obj.pk).language_code, 'en')
-            self.assertEqual(myadmin.get_object(get_request, obj.pk).translated_field, '')
+            if django.VERSION >= (1, 9):
+                self.assertIs(myadmin.get_object(get_request, obj.pk), None)
+            else:
+                self.assertEqual(myadmin.get_object(get_request, obj.pk).pk, obj.pk)
+                self.assertEqual(myadmin.get_object(get_request, obj.pk).shared_field, obj.shared_field)
+                self.assertEqual(myadmin.get_object(get_request, obj.pk).language_code, 'en')
+                self.assertEqual(myadmin.get_object(get_request, obj.pk).translated_field, '')
 
     def test_get_object_nonexisting(self):
         # In case the object doesnt exist, it should return None