+++ hvad/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '1.3.0'
-VERSION = (1, 3, 0)
+__version__ = '1.4.0'
+VERSION = (1, 4, 0)